Deleting, editing and retrieving data from the database full connection of gui with database

+++ src/main/java/exerciseCreator/Outcome.java
@@ -2,13 +2,13 @@ package exerciseCreator;
 
 public class Outcome {
 
-    private final float grade;
+    private final int grade;
     private final String firstName;
     private final String lastName;
     private final String studentData;
     private final String comments;
 
-    public Outcome(float grade, String firstName, String lastName, String studentData, String comments){
+    public Outcome(int grade, String firstName, String lastName, String studentData, String comments){
         this.grade = grade;
         this.firstName = firstName;
         this.lastName = lastName;
